+++ saleor/static/dashboard/js/components/components/cards/index.js
@@ -1,0 +1,9 @@
+import DescriptionCard from './descriptionCard';
+import FilterCard from './filterCard';
+import ListCard from './listCard';
+
+export {
+  DescriptionCard,
+  FilterCard,
+  ListCard
+}
